refactor: create streams module
